refactor(test-helpers): relocate to src/test-helpers.ts (out of legacy/core/)
Test mock factories (createMockSession, createMockSessionStore) for
transport unit tests. Was historically in src/core/ then dragged into
legacy/core/ during the #630 quarantine. Conceptually framework-wide
test utility — top-level next to paths.ts is the honest home.

- src/legacy/core/test-helpers.ts → src/test-helpers.ts
- 4 importers updated (slash-commands.test, http/chat.test,
  discord-manager.test, discord.test)
- File header + sessions/types relative import path corrected

Bottom-up cleanup of src/legacy/core/ — 1 file down, 4 to go.
