+++ src/legacy/commands/slash-commands.test.ts
@@ -2,7 +2,7 @@
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { SlashCommandHandler, type CommandContext } from "./slash-commands.js";
-import { createMockSessionStore } from "../core/test-helpers.js";
+import { createMockSessionStore } from "../../test-helpers.js";
 import type { AgentRuntime } from "../../agent/runtime.js";
 import type { AgentConfig } from "../../agent/types.js";
