+++ src/legacy/plugins/discord/discord.test.ts
@@ -4,7 +4,7 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { DiscordTransport } from "./discord.js";
 import type { SessionStore } from "../../../sessions/types.js";
 import { ThreadBindingManager } from "./thread-bindings.js";
-import { createMockSessionStore } from "../../core/test-helpers.js";
+import { createMockSessionStore } from "../../../test-helpers.js";
 import { AgentRuntime } from "../../../agent/runtime.js";
 
 function makeMockRuntime(agentId: string, cache: unknown, sessionStore: SessionStore): AgentRuntime {
