+++ src/legacy/plugins/http/chat.test.ts
@@ -6,7 +6,7 @@ import { ApiServer } from "./server.js";
 import { CronScheduler } from "../../automation/cron-job.js";
 import { SessionStoreManager } from "../../core/session-store-manager.js";
 import { AgentRuntime } from "../../../agent/runtime.js";
-import { createMockSessionStore } from "../../core/test-helpers.js";
+import { createMockSessionStore } from "../../../test-helpers.js";
 
 const MOCK_AGENT_ID = "mock";
